File location changes + import refactors
