+++ frontend/src/components/Beacon/InfoWindow/ControllerInfo.js
