+++ frontend/src/components/Beacon/MapContainer.js
